JS: Add hoverable previous/next to image viewer.
Adds previous/next hoverable panes, right above the image, to the text
viever. These previous/next buttons have the same behavior as the
current previous/next, but are easier to click than the small text,
making it useful for large images, large screens, or on mobile.

Change the JS behavior around the image to only delete the image instead
of the entire div, so that the buttons could be positionned absolutely
on top, in the same div as the image.