Images::Vote: Remove the 90% vote count limit
This is a bit of an experiment. Performance seems to be alright so far,
so let's see if we can bump it up a bit.

Downside is that people who've on voted almost all may sporadically get
no more new images. A fair trade-off, for now. If performance allows I
could add a 'just get all non-voted images and order by random()'
fallback.