v2rw/TraitEdit: Rewrite trait edit form + cleanup bunch of old code
This was the last form in the old v2 code, so this allowed me to get rid
of the old form validation and generation abstractions.

The form is mostly a duplicate of the tag editing form, as expected.
It's really annoying to have to maintain pretty much duplicate code for
tags and traits, but there are just enough differences that the
complexity of sharing code or creating a new abstraction outweigh the
benefits. :(