notifications: Consolidate and simplify creation of notifications
Preparing to extend the notification types, but with the old
individual-trigger-for-each-type that would easily turn into an
unmanagaeble mess. The new approach also correctly handles events that
may trigger multiple notification types.

Also removed the title and uid cache from the notifications table,
they don't simplify things and aren't necessary for performance. On the
other hand, doing a generic information lookup for vndbids is kind of
annoying. I now wrote an item_type() function to help with that, but
it's not sufficient to solve the broader problem.

One regression(?): Users will keep getting notifications for new posts
on threads even if they haven't read older notifications for the same
thread yet. I /think/ this behavior is more intuitive and expected, so
maybe I'll keep it.

(And possibly more regressions as well, who knows...)