reviews: Allow anonymous voting + display overrule counting
Anonymous votes count for 0.01, regular votes for 1 and overrule votes
for 1000. It's a bit crude, but ought to do the job.

Reason for this change is that most reviews don't seem to get any votes
at all. With this anonymous voting we can do some sort of
scoring/ordering even with potentially less reliable votes.