imgflag: Experiment with making average calculation more conservative
If an imgmod voted a more conservative option than the average, then the
imgmod's vote will be used as average. This is to catch cases where most
voters didn't notice a subtle suggestive or explicit part of the image.

I mean, if a mod votes more conservative, they must be on to something.