Advsearch: Initial experiments with a new advanced search
Doing this on the main branch to make it easier to get early testing and
feedback. Not like I have anything worth testing now, but it's not like
this code is getting in the way of anything else.

(Unless the changes broke something unrelated, in which case it's extra
good to get that early testing)