Advsearch: Change model to be suitable for advanced search
Turns out my initial idea with a seperate FQuery type wasn't going to
work out all that well, and it's easier to handle this by treating
nested And/Or queries as just another field. The nested field type is
going to be responsible for some extra tasks.

Also added a global 'Data' type to pass around in various functions.
This would be needed at some point anyway, but now that nested fields
can create/manage subfields, we'll need to be able pass around the
global object identifiers thing already.

And while I was fixing this model, I had a brilliant idea regarding the
UI. I was originally planning an explicit split between "quick selection
mode" and "advanced search mode", where you could only build nested
queries in the latter. But it's actually pretty trivial to make the
advanced search mode look like the quick selection mode in the simple
case. Buttons to change nesting type and to add more fields will result
in a mode natural flow between the two modes. I've not implemented the
UI for that yet, though.

...to be continued