imageflag: Add images.c_uids cache + remove tablesample on few images
This fixes spurious "No more images to vote on!" messages when the user
has already voted on almost all images and the tablesample doesn't catch
the few images that they hadn't voted on yet.

Since doing such a select without tablesample is *much* slower, this
needs a more efficient way of finding images the user hadn't voted on
yet, hence the c_uids cache. Still pretty inefficient - it needs a full
table scan - but it'll have to do.