AdvSearch: Fix using encoded form on pagination + abstraction changes
I was originally planning to have only two query encoding forms: Fully
normalized JSON form and compact encoded form. But converting between
the two is kind of annoying and requires lookup tables, which makes the
normalized JSON form inconvenient for Elm. Hence I added a third form:
compact JSON form, which has the same data type transformations applied
as the compact encoded form, but still retains the JSON encoding and
structure. Converting between compact JSON form and compact encoded form
is fairly trivial. Downside is that this also implies that Query
handling in Elm needs to be done through field numbers rather than
names, which is more error prone - but this can be solved by generating
an Elm module with a variable for each field numer. Another downside is
that this makes it impossible to implement a normalized-form query
viewer and editor in Elm without hitting the server for conversions -
but such a feature is not very important anyway.

Other abstraction change is that AdvSearch.pm now exposes an
object-oriented interface, the object can keep track of the different
query forms and seems like a more suitable solution in this case.