Advsearch: Add release date filter
This required some changes to the query model and encoded form, since it
doing this with only <= and => operators is too limited.

I changed the compact encoding of strings to only encode the length in
the type field for specific (common) lengths and fallback to an
end-of-string character for everything else. This leaves room for more
custom types.