AdvSearch: Add tag filters
The Tuple/Triple types really don't play well with
normalization/serialization at the moment. I had a more flexible type
system in mind, but it's impossible to write codecs for them without
explicitely encoding the data type in the JSON structure as well, which
is ugly...

Tuple type isn't used at the moment, but I have a few use cases for
that (traits and resolutions, perhaps staff as well, but that'll get
real ugly).