AdvSearch: Allow adding subquery filters directly from the "add filter" button
This changes the "add release filter and then add subfilter" action into
"select filter type then select the subfilter", which I suspect makes
the subquery feature more easy to discover and eases the construction of
complex queries.

I also removed the "And" / "Or" options from the list, as they're not
really part of the query type. Constructing complex nesting of and/or
chains is still possible through the arrow buttons on each filter
dropdown, but that's less intuitive. I'd like to re-add the And/Or
options to the add flter list somehow, but... differently. Not sure yet.