AdvSearch: Reorganize nested filters
Hopefully it's more intuitive this way, though to be honest I'm not
exactly sure it is. The switch between compact mode and expanded nesting
mode can be a bit surprising.

In terms of code organization, I added the and/or selection to the same
nesting node that handles the type difference. This provides better
control over how the queries are displayed, but removes the ability to
move the "and/or" itself into a parent node. Can be reimplemented if
useful.