Homepage: Merge & relayout some boxes
This change is likely to be a bit controversial, but I tried to keep the
links and boxes at roughly the same location: VNDB discussions where the
announcements used to be, VN discussions where the posts used to be and
latest reviews on the lower left.

The "Random VNs" box is gone, despite the negative feedback from last
time I removed it. It really is superfluous with the screenshots and
random VN button in the menu

And switch to CSS flexbox to layout the boxes, because for some reason
these changes triggered the not-full-width bug that also occurs on the
tags/traits indices.