AdvSearch: Add VN list/label filters
This is a bit tricky. The query includes the user id, so the URL can be
passed around to display the list of someone else (of course, taking
the private flag into account). The UI does not currently handle this,
so queries containing labels of other users will display an ugly
"Unrecognized query" box. Label filters without the uid can be
constructed as well, but again not through the UI and this will result
in an internal server error when a non-logged in visitor attempts to
open such a URL.