AdvSearch: Add staff role filter + fix bug with some OR queries
When the role filter is used as part of a Seiyuu subquery, it will match
the roles that the person has performed in any VN, not necessarily the
VNs where the character has been voiced. This is not ideal, I think, but
doing this any other way is likely going to kill performance.

The role filter when used as part of a VN->Staff subquery will match the
VN, as anything else would be fairly useless. This also has a few cases
where performance is impossibly slow and will timeout, but most queries
seem to work alright.