Convert images by running convert instead of loading Image::Magick
This has three potential advantages:
- Memory used for image processing is released immediately, thus
  resulting in lower average memory usage. I've yet to test how
  significant this is in practice... I'm not entirely sure how VNDB's
  memory use is distributed (not that it's really a problem, but if I
  see something that could be optimized I try to optimize it. First
  world engineering problems)
- The 'convert' command can be better sandboxed for increased security.
- Removes a dependency on a Perl module... except spritegen.pl still
  uses it, but that can be rewritten as well if necessary.

In the process I found out that the old code's attempt to sharpen the
image after resizing was buggy and the sharpening was never applied.
This new code *does* apply the sharpen filter, so newly uploaded images
will be visually different (if resized).

I also found that imagemagick's built-in ratio-preserving resize
sometimes ends up with different image dimensions than my imgsize(). I
think that's a bug in my imgsize function, but it does result in a
discrepancy that should prolly be resolved as it may result in
screenshot thumbnails being displayed incorrectly.