HTML/debug: debug text toggle + display interpolated SQL
Interpolated queries are a lot easier to quickly throw into the console
for debugging. The interpolation may be buggy, though, as values may not
be escaped properly or the wrong value is interpolated in the wrong
position (e.g. if there is a literal/escaped ? in the query or if
Postgres-specific $n variables are being used).