Image upload: Ignore return status of convert, just check if target file exists
I've disabled the 'XML' delegate library, which ImageMagick tries to
load when reading a jpeg file with XMP metadata. The image will be
processed correctly, but the return status of convert indicates an
error. For actual file loading errors, convert would error out before
writing the target file, so checking for that ought to be good enough.