v2rw: Some cleanup; Move some functionality out of vndb.pl; Move config out of tuwf->{..}
Config had been in a separate config->{..} object for a while now, old
code doesn't need it anymore so we can finally get rid of the tuwf->{..}
compact.

vndb.pl is not really the place for library code, so moved that around.
The old imgpath() and imgurl() functions weren't used anymore, so could
be safely replaced with the new functions that take the new vndbid
format.

(All this to make it easier to write separate scripts that call into
VNDB::/VNWeb:: code as a library)