AdvSearch: Only add quick-select fields when there are no non-quick fields
...and when the total number of fields doesn't exceed 4. I.e. the
quick-select fields are now only added if the entire query can be
described in terms of those quick-select fields, which means the user
probably never added or removed fields in the first place. This also
fixes the 'tag' field getting added even when it's already present.

Fixes the first point of https://vndb.org/t14995.55