Releases::List: Add experimental new release browser with advanced search
Uses the same release_row_() function used on VN and producer pages, so
this adds a lot more information and user list management functionality
to the release browser.

Removed compat with *very* old URLs, I'm monitoring the logs to see if
those are still used. Also removes a friendly message when the filters
don't match anything.

Main reason for implementing this before going live with the VN advanced
search is to make it possible to also switch the default saved release
filters to the new system in one go.