AdvSearch: Change add filter menu structure
Re-adding the listing of "And/Or" subqueries. Nested filters are now
part of the regular filter listing and open a submenu with related
filters for that type. This way the menu structure allows arbitrary
nesting and follows the same structure as the overall query itself.
This should also make nested queries more discoverable.