Homepage: Fix creating an invalid releases query with empty branch
The old filter->advsearch conversion would result in some people having
the following query as default release filter:

  [ 'and', [ 'released', .. ] ]

The homepage code would get rid of the 'released' filter and end up with
an empty [ 'and' ], which gets encoded into an invalid query string.
This fix removes the branch if it's empty.