elm/RDate: Fix date selector to adjust the "day" field when date would be invalid
It used to be possible to construct invalid dates by first selecting a
full date (e.g. 2021-01-31) and then changing the month field so you'll
get 2021-02-31. Or going from 2020-02-29 to 2021-02-29. I've adjusted
the normalize function to catch these invalid cases and correct the day
part to the correct last day in the month.