v2rw: Convert VN releases tab
This is the last page in the old v2 code that hadn't been rewritten yet.
It's a fairly direct and ugly conversion as I've not decided what the
future holds for this page, but at least this will allow me to clean up
a lot of old and ugly code.