Add function to lookup user by email address
Not a huge fan of the user_emailtoid() SQL function as that makes it
even easier for an attacker to check VNDB accounts against an list, but
this was already possible with the password reset functionality anyway.
At least there's no index on the email column, so any kind of brute
force search will show up soon enough in my monitoring.

The email lookup feature is needed so I can answer account support
requests coming in from email without having to query the DB manually.