cleanup: Merge VNDBUtil into VNDB::Func
Something I was planning to do in the v2rw rewrite. VNDB::Func used to
be a module intended for use inside TUWF (i.e. the web backend) while
VNDBUtil was for utility functions that could also be used outside of
that. Now that the web backend has moved to the VNWeb::* namespace,
the VNDB::Func module has taken over the role of VNDBUtil.

Also updated the API code to make use of the 'imgurl' function, which
is a nice additional cleanup.