SQL: Improved column packing + reformat schema.sql
This only saves about 1% of the database size on the production DB, so
it's rather insignificant. But hey, the original column order didn't
make sense, nothing in the code (should) depend on column order, and I'm
about to re-import the database anyway. Might as well get a little
cleanup and efficiency boost while I'm at it.