SQL: Reduce precision of cached floats + avoid unnecessary batch updates
Avoiding UPDATEs to rows that don't need to be updated should
significantly reduce table bloat and improve the performance of the
batch queries by a bit.

Rounding numbers to the precision that we actually need helps to avoid
unnecessary row updates and should also reduce the size of database
dumps by a tiny bit, assuming floating point formatting won't ruin it.