DB::db_entry(): Reduce query count + fetch from non-_hist tables
I merged three queries into one larger and more complex query, but
planning and execution time of that query is fast enough to make that
worthwile.

The most common case of db_entry() is to fetch the latest revision, so
it makes sense to optimize that by fetching from the base tables instead
of the history tables - the base tables are more likely to be in cache
and have fewer rows and thus smaller indices. Overall, though, this is a
very marginal improvement, considering the entire DB easily fits into
memory anyway.