HTML: Use hiddenmsg for tags/traits and to display awaiting moderation state
This also fixes the issue that the reason for deletion is not displayed
if it's in the message of the last change, which is the case for newly
rejected tags/traits. The edit message is not displayed if the entry was
deleted in the first revision, as that's most likely an import.