SQL: Expand & improve item_info() and use it where it makes sense
The biggest performance issue with the previous implementation of
item_info() was that PostgreSQL would generate a query plan for each
invokation, and planning such a long query takes time. This new
implementation improves on that in two ways:
- Using much shorter queries that can be planned faster, using PL/pgSQL
  to select the proper query.
- PL/pgSQL can cache query plans for queries that are executed often, so
  for listings and longer sessions there is no planning overhead at all.

These optimizations make it fast enough to use item_info() in places
that would previously maintain separate query lists, UNIONs or JOINs for
each entry type. In some cases this approach is even faster.
Yay, code re-use!