A few email delivery related changes
Make it clear that a valid email is required in order to create an
account. The previous wording implied that there was no account
activation and people would enter bogus addresses.

There (should) not be a need anymore for the "shitty email provider"
messages, as I'm routing mails for those destinations through external
relays now. I really hate that this is necessary, but such is life.

Also fix the wording of the password reset mail - that function hasn't
disabled your account for a while now.