Ulist.Widget: Experiment with a VN list management widget
Currently only added on VN listings in row view, but the goal is to add
this widget to other listings and pages as well. This should make it
easy to see whether a VN is on your list and to add/remove/edit your
list status.

The code is one of the messiest copy-paste jobs I've ever done. I'm
totally going to regret having to maintain this crap. But anyway, let's
first see how this UI is received.