AdvSearch: Replace "developer-id"/"producer-id" fields with producer subfilters
Backwards compatibility with old filters is surprisingly tricky. I was
originally planning to convert old queries to use these new subfilters,
but handling multi-producer-selection is nontrivial. There is no
equivalent for the "releases.developer-id" field now, but I did some
logging and it doesn't seem to be used anyway.

My current approach to backwards compat is to just keep the old filters
around, both in the backend and in Elm, but simply hide them from the
filter menu. Not sure this is a good long-term strategy.

Anyway, this makes it possible to filter by producer types, as in
https://vndb.org/t15372.107