SQL: Move utility functions to separate util.sql
This is a cleaner solution to the problem of the schema relying on a few
utility functions and creating a dependency cycle between schema.sql and
func.sql.

Also fix dbdump.pl to not export generated columns.

These changes are to prepare for an upgrade to PostgreSQL 14 with a
clean import through dbdump.pl. Which appears to work surprisingly well
in my testing so far.