Move & streamline VN release expanding settings
Problems with the old implementation were that the setting wasn't an
intentional user action, so there were a LOT of preference writes that
were reverted a minute or so later. Wasting both server and visitor
time. For regular visitors, the localstorage-based approach also caused
a reflow if their preferences didn't match the default.

Downside of the new implementation is that regular visitors can't change
their preference at all anymore, this now requires an account. :/