Add tags_vn_direct cache table
This speeds up VN pages a little bit, but the main reason for this table
is to implement no-parent-inherit tag search.

Because the rows in tags_vn_direct represent an intermediate calculation
for tags_vn_inherit, generating and updating these caches isn't all that
much slower with this change. I'm not a fan of the DELETE & re-INSERT
approach on individual-VN updates, though, as that tends to cause
unnecessary (index) writes. Not sure how complex a more incremental
update is to implement, but it sounds painful. :/