SQL: Reduce size of images table + set a max of 15 votes
Reduction is by 12 bytes per row, by simply encoding the same data as
smallints. We never needed the full range of the real/integer types.
(Could go even further, but postgres doesn't have a uint8 type and
encoding avg + stddev in the same column feels a bit too hacky)

Having more than 15 votes on an image is pretty useless, so by removing
images with sufficient votes from the image voting selection we can
avoid further bloating the table.

(c_weight is still >0 even for images that shouldn't be selected
anymore, because this field is also abused to check if an image is
referenced somewhere in the DB. Instead, images with c_weight=1 are now
excluded from selection. This has the side effect of also excluding
images with 14 votes and stddev=0, but that seems okay)