Consolidate UList tableopts with VN::List
Goal is to slowly consolidate more UList code with VN::List, so that all
columns and display options of the latter are also available on the
former. The only real change with the current commit is that more
columns can be selected in UList, but this is just step one.