SQL: Experiment with MERGE to perform in-place updates for tags_vn_calc()
The previous TRUNCATE + INSERT approach is still faster, but has the
major downside that it acquires an exclusive lock on the cache tables
and hence brings down the site for 10 seconds. This MERGE approach is a
bit slower, but avoids unnecessary writes and the site should hopefully
remain responsive. Both implementations perform equally well for
single-VN updates.

If this approach turns out to work, I'll also fix traits_chars_calc(),
which is responsible for taking down the site for 20 seconds every day.

If this doesn't work, new table + rename is prolly also a viable option.