SQL: Also use MERGE for traits_chars_calc()
Commit d073490d4745fbc5db751fc15135311d803f1d66 worked out very well, so
expanding the use of MERGE to the traits cache as well.

There was no index on traits_chars.cid, so incremental updates were
pretty slow (150ms+) with either approach. Added an index to bring that
down to 15ms. One advantage of this MERGE strategy is that indices are a
bit cheaper; after all, avoiding writes also means less index updates.