TitlePrefs: Also apply preferences to producer names
Bit of an experiment. I should probably swap the schema columns too
(i.e. name + latin) to be consistent with the recent VNs & releases, but
that's not really a requirement for applying these preferences.

And the more places where title preferences are applied, the more
important it becomes to also set proper html "lang" attributes, which
this commit doesn't do either.