Rework passing around entry titles + add proper HTML5 'lang' attrs
This is a huge back-end change for a small benefit, but it also improves
consistency a bit in the way that titles are handled and passed around.
The Elm code still uses separate title/alttitle fields at various
places, not sure if I want to bother converting that, too.

I'm sure there's bugs, given how many files this touches.