Exclude unnecessary hidden DB entries from search cache
It was possible to search for deleted producers, but I didn't really see
the value in that. Checking for deleted VNs is still important, likewise
for tags & traits.

This reduces the size of the cache a bit, but more importantly also
improves accuracy of the search result counts for other DB entries.