CSS: Apply default reset for box-sizing: border-box
This is part 1 of a probably large set of changes to improve VNDB's CSS
mess. border-box seems more intuitive to me and is a common CSS reset.

I didn't go for a pixel-perfect conversion, so this changes the
dimensions of a whole lot of elements on the site. Didn't stumble upon
too many serious issues though, and in fact found at least two cases
where this *fixes* a misalignment.