SVG: Optimize and merge rss & list icons
The idea behind the list icon CSS filters was that they could be
adjusted in themes, but (1) that never happened and (2) even with colors
embedded in the SVG they can still be adjusted with the same mechanism.
In that light, embedding the colors in the SVG is the cleaner solution.