CSS: Now with more grid and less positioning
Moving the top-level maintabs into the header, so that the Menu button
nicely fits into it. This fixes a wrapping problem when the page tabs
are too wide.

Also moved some margins and paddings around again, so that there's much
less calculated values that are dependent on each other. This should
simplify customization and mobile adjustments quite a bit.

Also fixed a z-index problem with #bgright.