UserEdit rewrite part 5: Public traits + some dropdown search features
Not yet super happy with the table-in-form layout thingy; buttons jump
around when adding/removing items, ugh.

I *am* quite happy with the dropdown search thingy, the new features
make it much more intuitive to have items that can't be selected. The
current Elm code usually ignores the selection of items that are already
present or otherwise unselectable, which can be rather confusing.

Having the input field replaced with a button is something I'm not yet
decided on. It does makes sense, but also requires an additional action
when adding multiple traits. Even if that action is just pressing the
enter key. Multi-select built into the DS abstraction might be a
solution.