Producer::Edit: Convert form to JS
This one got priority because many people appear to be stumbling on the
URL validation bug in this form. Staff edit form is still affected by
it, but that's less commonly a problem.

New form also doesn't pre-fill a language anymore and instead requires
the user to input one manually. Fixes a problem with non-JP producer
entries having JP selected by default.