Build: Drop zopfli(png)
For CSS & JS, brotli now serves 95% of all requests. zopfli saves a few
bytes for the remaining 5% that still use gzip, but it slows down my
builds quite a bit and that's not really worth it anymore. gzip -9 is
good enough.

For icons.png, zopflipng made a lot of sense when the platform icons
were still included. Nowadays the icons.png file is already so small
that further compression with zopflipng doesn't save much.