JS/VNLengthVote: Rewrite form, use separate page + minor improvements
The separate page might be considered a slight regression, as this
increases friction a little bit, but the extra space this gives us does
make the form and the various available options less confusing. I hope.

If the friction is a problem, the form could also be kept on the VN
pages in a popup-style fashion, but that doesn't have my preference.

Also adds support for voting on trails for cancelled VNs.

(Also paves way for re-using the form for mod editing, but that's not
too important, current mod options seem sufficient)

(Also damn this form is a lot more complex than I had expected, so many
edge cases)