Enforce maximum size of images after re-encoding
Context: https://vndb.org/t14195.241

Ideally a re-encoded image doesn't increase in size, but this can happen
when the original upload format isn't JPEG or with different quality
settings. Would be nice to still keep the image in the latter case, but
that's tricky to do while stripping metadata and ensuring the image
doesn't have a badly supported color profile or something.