SQL/chflags: Support excluding columns and setting multiple flags per column
Fixes two issues. The 'docs.html' column should not be considered for
changes since it might be slightly different after an update to the
markdown convertor.

The VN/release languages should also be considered a "Title" change,
because otherwise swapping the titles between two languages might not be
considered a "change".