SQL: Move release/staff external links into a separate 'extlinks' table
The immediate motivation for this change is to debloat the releases &
staff tables in order to improve query performance.

But that could also be achieved by moving all the link columns into a
separate table, which would be much less invasive. The other reason for
this change is to build a framework for an automatic link checker.
Whether that will actually happen, though, I'm not sure, because the
main challenge for that is to figure out how to deal with anti-crawling
measures that every site has nowadays. But anyway, this new schema is
also useful for the few sites that we already crawl: the affiliate
links.

The new extlinks form also automatically copies matched URLs from the
website link when editing, which should prevent some typical errors.

VN and producer entries also have a few external links that I haven't
converted yet, so there's a bit of legacy code left over. I'll migrate
those eventually. Those entries deserve more links as well.