Tighten security around usermod features
I realized that blind SQL injection on a usermod account could be used
to gain access to a(nother) usermod account, which could then be used to
list everyone's email addresses. A fairly obscure attack vector and even
with these changes such an SQL injection could do a lot of damage, but
the impact is slightly lessened.

In particular:
- usermods can no longer set the usermod flag, this has to be done
  through SQL.
- usermods can no longer change the email address of another usermod.
- changing email address of a usermod has to go through the usual
  confirmation flow now.
- while this functionality was not exposed through the web UI, usermods
  can now also not flag another account for deletion or change their
  password.