SQL: Add vndbtag type and rework vndbid type
Improvements compared to the old vndbid:
- Much wider range of supported types ([a-z]{1,3}) instead of a
  limited predefined list).
- Support for much larger numbers (48 bits instead of 26)
- Support for type modifiers to restrict the type part
- Convenient index-capable prefix match operator (id ^= 'v')
- Convenient operators to extract type or number
- A separate type to efficiently represent the tag
- Proper tests \o/

The downside: I had to widen the type to occupy 64 bits instead of 32,
so tables are going to be a bit larger and less efficient now. I'll need
to do benchmarks, but don't expect this to have a significant impact.

Migrating an existing database to the new type requires a full dump &
reload through util/dbdump.pl. But no need to do that yet - I plan to
make a few other changes while I'm at it.