UList: Rewrite widget into JS and break everything else
Some slight UI changes to the widget itself, including:
- Different dropdown input for label selection, votes & releases
- Uses the release date input for start/finish dates
- Newly created labels are now immediately visible to other VN widgets
- Fixed CSS scrolling on overflow

This change completely breaks the list tool on VN pages and all inputs
on the user list in the table view. The idea is to re-implement those in
terms of this new global widget while sharing the same VN object
whenever possible, so that all inputs have a consistent view of the
current state.