UList: Fix some conflicts when editing labels from multiple tabs
Adding a label in one tab and then adding it again with the label editor
would result in a constraint validation error; now it correctly applies
the latest settings to the previously added label.

This also fixes a few other conflicts around removing and renaming or
adding a label with the same name.