FU: Convert API, search & AdvSearch queries
This is a larger change than I'd prefer, but these queries are all
interwoven through the AdvSearch & SearchQuery abstractions, so I
couldn't easily do a partial migration. I'm pretty sure this commit
breaks a thing or two.

On the upside, this is the final part of the migration to FU. All
remaining use of SQL::Interp has been converted now.

The API still uses the old enrich() conventions, I could convert those
to use fu->enrich() now that we have a more flexible function.