Add basic edit conflict detection & prevention
It sucks that this causes the second edit to be lost, but that's still
better than letting it go through and override the other user's edit.

In an ideal world the conflicting edits would get automatically merged
as far as possible and the user only has to adjust the conflicting
fields, but that sounds like a huge pain to implemenent.