Task: Convert unusedimages.pl and schedule it weekly
I had already been running that from a custom cron script, much simpler
to just merge it into the task queue.

It does use a lot of RAM due to keeping a hash table with the id of
every image in the DB, I'll have to see if that ever gets reclaimed in a
longer-running process.