Task: Merge daily tasks
Just realized that using slight increments in align_add is not a good
idea for sequencing tasks. The whole task executor is designed so that
it's possible to run multiple processes in parallel, which means tasks
scheduled near each other may run in parallel as well. That doesn't make
a whole lot of sense for the database maintenance queries.

But we can still enjoy the ability to run individual queries with the
CLI argument thing. \o/