SQL: Remove SKIP LOCKED from update_extlinks_cache() task update
Did that to prevent the function from blocking while the task is running
and assumed it wasn't necessary because the task's nextrun is calculated
afterwards anyway. Except the nextrun query might not see the row we
just updated, so we still need the lock to ensure the task will run
again soon. It was a silly optimization anyway, the el-triage task is
pretty fast.